client + server protocol cleanup
