+++ Sources/Vapor/Engine/SecurityLayer.swift
@@ -1,0 +1,6 @@
+import TLS
+
+public enum SecurityLayer {
+    case none
+    case tls(TLS.Context)
+}
